feat(source-map-config-issues): Iterating
